emove ui8 with updated XUSG, since it is not standard C++.

+++ MeshShader/Content/Renderer.h
@@ -30,7 +30,7 @@ public:
 	void Render(XUSG::Ultimate::CommandList* pCommandList, uint32_t frameIndex,
 		const XUSG::Descriptor& rtv, PipelineType pipelineType);
 
-	static const uint32_t FrameCount = 3;
+	static const uint8_t FrameCount = 3;
 
 protected:
 	enum PipelineLayoutIndex : uint8_t
